Ground Control Station/API/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.4
python-multipart==0.0.6
pyserial==3.5
requests==2.31.0